Componentize AutoLoginPrompter::ParseAutoLoginHeader
The reason is to re-use the x-auto-login parsing logic in
android_webview/.

The AutoLoginPrompter::ParseAutoLoginHeader is moved and the header
parts of the Param struct is extract into a component. The unit_test has
also moved as they are only testing the parsing logic, but still using
the chrome target.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11970026