Have create_nmf.py generate a hash of PNaCl pexes for caching.
Without it, pnacl doesn't cache any translations which means
that pnacl browser_tests won't be able to test caching.
It also means users won't benefit from caching.

Alternatively, we could use a sha embedded within the pexe
that the pexe linker generates, but we have not converged
on which is better...

sort of related to
BUG=http://code.google.com/p/nativeclient/issues/detail?id=2992

Review URL: https://chromiumcodereview.appspot.com/12017030