Enable building of test plugin on ARM.
The test plugin is needed by several browser and content
tests.  The missing plugin was causing these to fail
on the ARM linux bot.   This change enables the plugin
for ARM in general (e.g. chromeos and linux) while leaving
it disabled for android and ios builds.

BUG=169810,169813

Review URL: https://chromiumcodereview.appspot.com/11931003