cc: Use the size of the source frame along with the source frame's renderpasses
Scale the contents of the source frame to fit exactly in the bounds of the
delegated renderer layer.

When the host window is being resized, it can keep the delegated renderer layer
at the size it wants the output displayed, and clip it with a containing layer
until the resize is complete and it has a new frame to show at the right size.

Tests:
DelegatedRendererLayerImplTestSharedData.SharedData
DelegatedRendererLayerImplTestSharedData.DisplaySize

BUG=123444
R=enne
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/11821015