Add hack that works around signaling bug that breaks Chromoting.
Signaling endpoint used by chromoting doesn't always handle correctly the
case when host and client use the same stanza ID's. Workaround this issue
by adding 'x' prefix to all messages that go from client to host.

This is the same change as in crrev.com/160330 . That CL was reverted 
because it was believed that the original bug was fixed.

BUG=168617


Review URL: https://chromiumcodereview.appspot.com/11926031