Limit status at 500.
Otherwise Status().put() raises an exception and an HTTP 500 is returned to the
user.  Very long messages can happen when gatekeeper closes the tree and there
is a very long blame list. It's better to cut is short than to not close the
tree at all.

R=csharp@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/12017005