Omnibox: Control-Key Interaction with Keyword Search Mode
Remove two lines that appear to be unnecessary.

These lines were introduced more than two years ago in this change:
https://codereview.chromium.org/3322001/
To fix these bugs:
http://code.google.com/p/chromium/issues/detail?id=44591
http://code.google.com/p/chromium/issues/detail?id=18431

I tested this by playing around as much as I could using the control key
and keyword mode with current Chrome and with Chrome with this change.
I didn't notice any changes whatsover.

I also tried to reproduce the first bug.  It's not exactly reproducible
as it's described (in comment 3) anymore.  Previously it appears that
by selecting a suggestion with a keyword prompt would automatically put you
in keyword mode (with the tile: Search blah for:).  This doesn't happen
anymore; when you arrow to select a suggestion with a keyword hint, you
see the URL for the site and a prompt "press tab to search".  You don't
automatically go into keyword mode.

This change leaves the current behavior unchanged.

The only affect of this change appears to fixing the DCHECK bug linked below.

BUG=145185


Review URL: https://chromiumcodereview.appspot.com/12033013