Temporarily disabling WebRTC video detection on Linux, changing back video algorithm settings.
The background is that we have had problems with flakes a long time when running these tests in a Release-built chrome. I've investigated it extensively and the only possible reason I can see is that it's either an extremely exotic race condition or dependent on x environment or machine the tests are running on. These tests have looked more stable when running in Chrome infra, so I'm disabling this until we move there (which should be shortly).


TBR=craigdh@chromium.org
TESTED=Locally, and on a real build slave.
BUG=165135
NOTRY=True


Review URL: https://chromiumcodereview.appspot.com/11896032