Implements onReceivedIcon
This patch implements both WebChromeClient.onReceivedIcon and WebView.getFavIcon, and provides most the  support for WebChromeClient.onReceivedTouchIcon.

It also includes instrumentation tests for onReceivedIcon.

What is still missing:

Ability to turn it on and off.
TouchIcon currently isn't working.
Navigation history is not updated as of this patch.

BUG=

Review URL: https://chromiumcodereview.appspot.com/11575022