Fix unsafe use of WeakPointerFactory in BrowserOptionsHandler.
The previous code was using two weak pointer factories,
detaching one so that it could get bound to the FILE thread
for the "this" parameter of CheckAutoLaunch().

However, that method didn't use any instance state, so this CL
changes it to be static and removes one of the factories.

BUG=171329
TEST=no functional changes

Review URL: https://chromiumcodereview.appspot.com/12042016