cc: add workaround to change text color of the FontAtlas
This change adds the method FontAtlas::setColor() and the helper class
ScopedFontColorFilter, which make it possible to change the HudLayer's text
color. The workaround using a color filter is necessary because the FontAtlas
uses a bitmap with prerendered glyphs for text drawing.

In addition this change fixes the text colors on the HudLayer and adds
new colors to the DebugColors.

BUG=


Review URL: https://chromiumcodereview.appspot.com/12016025