Increased notification close button click area & added state images.
Expanded the clickable area for notifications from 8 x 8 pixels to
29 x 29 pixels and added hover and pressed state images.

Implementing this required moving the close button from a grid layout
column to a separate layer of the notification view, which caused a
number of changes in NotificationView's drawing.

In passing, references to one of the notification types the notification
view supports were changed from "multiple-item notifications" to the new
(and improved :-) "list notifications".

BUG=168822


Review URL: https://chromiumcodereview.appspot.com/11946035