Fix error breaking Chrome Endure tests, and update some tests for UI changes.
An error was introduced in https://chromiumcodereview.appspot.com/11740020
that broke the Chrome Endure tests (due to a reference to an undefined "self").
This CL fixes this error, and also makes a few changes to some of the Chrome
Endure tests to account for recent UI changes on the live sites.

BUG=None
TEST=Verified the modified tests now work with these changes against the
live sites.
NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/12052031