Add app_list_unittests as a default builder target.
Prior to this change, app_list_unittests was built only on aura targets.
However, the app list has been ported to Windows, and a mac port is
in progress. Once this change bakes, a trybot target will be added for
the builders in the tools repo. app_list_unittests is currently small and
quick (runs in <10 seconds).

BUG=138633
TEST=builders succeed their compile phase

Review URL: https://chromiumcodereview.appspot.com/11820019