[chromedriver] Upload builds and handle Chrome/ChromeDriver versions.
This adds run_buildbot_steps.py, which will be called by buildbot
instead of run_all_tests.py. The new script invokes run_all_tests.py and
uploads the ChromeDriver binary to googlecode if it's a unreleased
version and if all the tests pass. Thus, to release ChromeDriver, all
that is needed is to increment the ChromeDriver version.

Also, don't allow ChromeDriver to control an incompatible Chrome version.
As part of getting the Chrome version, fix ChromeImpl::EvaluateScript
to return the pure script result and not expect a {status:, value:}
like ChromeImpl::CallFunction should.
BUG=169087

Review URL: https://codereview.chromium.org/12052004