Extensions that use WebRequest API should not interact with <webview>
This patch exposes the flag that indicates whether a render process is a guest
to the IO thread in the browser process. When the WebRequest API checks for
matching listeners, it checks if the request is coming from a <webview> guest
process and if so, it checks if the listeners have a target process/routing ID
(no listeners currently do). Thus, the current WebRequest API no longer
interacts with <webview>.

Upcoming patches will enable the WebRequest API for guests via a slightly
different API.

BUG=171421
Test=Manually

Review URL: https://codereview.chromium.org/12052014