Wrote WebRTC audio detection test.
Note: I would have preferred to write this as a browser test, but the WebRTC browser test is too bare-bones at the moment; it needs a lot of work before it's mature enough to write this kind of test on. This will get something in place for now.

TESTED=locally, with and without correct microphone config on the machine. Also tested on the bot it will eventually run on.
BUG=
NOTRY=True


Review URL: https://chromiumcodereview.appspot.com/12052041