Mac 64-bit compatibility for remoting: use proper types.
DisconnectWindow is an NSWindow. NSWindow defines an instance method
-initWithContentRect:styleMask:backing:defer: where the type of styleMask
is NSUInteger. DisconnectWindow implements this method by using an unsigned
int as the type of its styleMask parameter instead. In the 32-bit environment,
NSUInteger is an unsigned int, and so the types are compatible. In the 64-bit
environment, NSUInteger is an unsigned long instead of an unsigned int,
resulting in this compilation error:

remoting/host/disconnect_window_mac.mm:181:41: error:
    conflicting parameter types in implementation of
    'initWithContentRect:styleMask:backing:defer:':
    'NSUInteger' (aka 'unsigned long') vs 'unsigned int'
    [-Werror,-Wmismatched-parameter-types]
                styleMask:(unsigned int)aStyle
                           ~~~~~~~~     ^
.../SDKs/MacOSX10.6.sdk/System/Library/Frameworks/AppKit.framework/Headers/NSWindow.h:264:69:
    note: previous definition is here
- (id)initWithContentRect:(NSRect)contentRect styleMask:(NSUInteger)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag;
                                                         ~~~~~~~~~~ ^
Review URL: https://codereview.chromium.org/12042089