Add support in NaClBrowserTests for PNaCl, and test PNaCl UMA.
Mostly the same as the previous CL from dschuff:
https://chromiumcodereview.appspot.com/11348048

Updated to only build pexes, and test the in-browser translation
of the pexe instead of pre-translating the pexes
(so create_nmf calls are different, etc.).

Add some UMA tests for PNaCl while we are at it.

NOTE: Tests are disabled under ASAN.
BUG=https://code.google.com/p/chromium/issues/detail?id=171810
In a later CL we can enable a test for pnacl_coordinator
under asan, but using incognito mode, which avoids the
caching logic.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2484
BUG= http://code.google.com/p/nativeclient/issues/detail?id=3136

Review URL: https://chromiumcodereview.appspot.com/11956028