Stop proxying transactions through onUpgradeNeeded
This is a small thread of code I need to remove that allows me to refactor a bunch of
WebKit code. The transaction makes its way all the way through to IDBCallbacks::onUpgradeNeeded
but nobody actually uses the transaction itself. This removes the proxying in the smallest manner.

BUG=


Review URL: https://chromiumcodereview.appspot.com/12033085