Avoid going through WebWidget::composite to composite a RenderWidget
When a content::RenderWidget is using accelerated compositing, the compositing
is driven either by a WebLayerTreeView owned by the RenderWidget or a PepperWidget
if the RenderWidget is actually a RenderWidgetFullscreenPepper. Neither case
involves WebKit, so this avoids calling WebWidget::composite in favor of directly
calling in to the appropriate compositing system.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11787006