Change ChromiumWebRTC waterfall to trigger on Chromium revisions
We'd like the master.chromium.webrtc waterfall to be 
triggered by Chromium revisions instead of WebRTC checkins.

The reason for this is that we want to build as frequently as
possible so we can more easily narrow down the interval of
Chromium revision when a breakage in WebRTC functionality has
occurred.

BUG=none
TEST=started local master and waited for a revision to trigger a build.


Review URL: https://chromiumcodereview.appspot.com/12047083