Introduces 'context' param to NotifierSettingsView.
I don't know why the existing code was once running on my local environment,
but 'context' parameter in WidgetParams is mandatory in Aura. To specify
the context properly, the best way would be to add another method to show
the settings window with the current context.

BUG=None


Review URL: https://chromiumcodereview.appspot.com/12052057