Declarative Web Request: firstPartyForCookiesUrl
This adds a new URLMatcher attribute to declarative Web Request conditions.

Because it looks like no other URLMatcher attribute will be added in the near future, this is done by hard-coding two URLMatchers in the code, instead of having a generic collection of them.

Unit tests only check construction. Because the rest of the code is shared with the request URL matcher ('url' attribute), there is no need to test again that, e.g., hostEquals matches are done properly.

An API test is used to verify that 'firstPartyForCookiesUrl' works and is different from 'url'.

BUG=159576

Review URL: https://chromiumcodereview.appspot.com/11414230