Use a more flexible methord to search for hashed names in webgl program log.
Before we hardwired the method to look for "webgl_" and next 16 characters.  Now we look for "webgl_" and the end of next hex digits.  This is less dependent on the ANGLE implementation than before.

However, to be on the safe size, still DCHECK if we found 16 hex digits.

BUG=160369
TEST=
TBR=gman
Review URL: https://codereview.chromium.org/12038103