cc: Sync tilings from pending to active tree.
We were previously assuming that the active tree always created tilings
and synced to the pending tree. However that's problematic because any tree
can be dirtied and have ManageTiles called on it. On commit, the pending tree
will make tilings in particular. So, sync tilings from the pending tree to the
active tree when adding to the pending tree.

BUG=172291
R=enne


Review URL: https://chromiumcodereview.appspot.com/12047120