Track UMA stats for when the renderer side audio device wasn't ready.
A renderer that's not ready causes glitching, we should measure what
percentage of users are having issues to determine how big of a problem
this is.  To start, we'll record the percentage of clients on each
platform who see more than 3 instances of an unprepared renderer.

BUG=none
TEST=<in progress> Run locally on each platform to ensure common case
is zero.


Review URL: https://chromiumcodereview.appspot.com/11975031