Remove ".." from PathService take #2.
Hook directly into PathService::Get() and only convert to an absolute path when
".." is found in the path. Otherwise leave it as-is. In particular, this won't
trigger when the path is simply relative, like "./unit_tests". This is
necessary to make most test cases pass but this has the advatange of not
triggering in the common case for Google Chrome.

R=cpu@chromium.org,mark@chromium.org
BUG=168890


Review URL: https://chromiumcodereview.appspot.com/12090003