cc: Throttle tile priority updates to once a frame.
The tile priority updates happen when calcDrawProperties is called. Now we
make sure we early out except in the following cases.

On the pending tree, we calculate tile priorities:
- Immediately after a commit, so that we can start raster jobs appropriately.
- On an animation tick if it is not the same impl frame as the commit occured
in.

On the active tree, we calculate tile priorities:
- During draw, if it is not in the same impl frame as the tree activation
occured in. In that frame, we already have priorities from the pending tree.

BUG=170807


Review URL: https://chromiumcodereview.appspot.com/12045086