Browser Plugin: Remove guest renderer flag
The --guest-renderer is no longer used by any BrowserPlugin code.

This patch removes the flag completely. This makes the Browser Plugin
architecture more obvious: a guest renderer is just another renderer
that is oblivous to the fact that it's being hosted within a <webview>.

BUG=none
Test=WebViewTest.*, BrowserPluginHostTest.*

Review URL: https://chromiumcodereview.appspot.com/12096010