Add sign-in promo to the Extension Installed bubble (mac).
Overview of xib changes (see screenshot in bug):
Added a TextField placeholder for a sync sign-in promo.
Widened the TextFields just a bit as there looked like there was a bit too much whitespace to the right of them (compared to other sides of the bubble).

BUG=164955
TEST=Automated tests.
Review URL: https://codereview.chromium.org/11997002