views: Fix a Widget destruction crash.
For WIDGET_OWNS_NATIVE_WIDGET widget, we destroy root view before we delete
|native_widget_|. If a WidgetDelegateView is in root view's hierarchy, the
native widget destruction would crash because |widget_delegate_| is accessed
in OnNativeWidgetDestroying but it is deleted with root view already.

This CL marks WidgetDelegateView as owned by client so that it does not go
away with the root view hierarchy and be deleted on DeleteDelegate.

BUG=164791
TEST=Covered with a new unit test.

R=ben@chromium.org


Review URL: https://chromiumcodereview.appspot.com/11953066