Merge 178906 - H.264 videos would stop playing randomly on Windows 8.
On Windows 8, the video decoder returns the MF_E_NOTACCEPTING error when we pass an input sample
to it for decoding. This error means that the decoder has enough data to produce one or more out
samples. We already had code for handling this, i.e. attempt to generate an output sample and
call ProcessInput again. This apparently is not enough as the decoder returns MF_E_NOTEXPECTING from
the second ProcessInput call which probably means that we need to process multiple output samples before
calling ProcessInput again.

Fix is to add the input buffer to the pending input list if we receive the MF_E_NOTACCEPTING error the
second time.

To ensure that we don't map and unmap multiple times for an input buffer, we convert the input buffer to 
an IMFSample upfront and keep the interface pointer in the pending input buffers vector along with the
buffer id.

Fixes bug https://code.google.com/p/chromium/issues/detail?id=164801

BUG=164801
Review URL: https://codereview.chromium.org/12049054

TBR=ananta@chromium.org
Review URL: https://chromiumcodereview.appspot.com/12087041