cc: Add a difference iterator to TilingData
This iterator lets one walk through all the tiles in one rect that are not
intersected by another rect.  The purpose of this will be to allow you to use a
"last frame rect" as the ignore rect and be able to only update tiles that
won't be updated this frame but were updated last frame.

R=danakj@chromium.org
BUG=171826


Review URL: https://chromiumcodereview.appspot.com/12096005