Add Brand Code to UMA Uploads
Can you suggest a good way to test this?  I don't have a Windows machine
(these are the only types of installs with brand codes).  Or do we
simply trust the code enough that we should just check this in and
see if after the next dev release that we get a person or two?  (There
probably are occasional new installs of chrome dev channels.)

BUG=


Review URL: https://chromiumcodereview.appspot.com/12036047