Rename chrome_frame coverage_bundles.py target.
chrome_frame and chrome use the same mechanism for generating a list of coverage targets. The chrome_frame targets have been prefixed with gcf_ to make them unique, but this was not done for the coverage_bundles.py target. This causes ninja to complain loudly that a target has multiple definitions.

BUG=


Review URL: https://chromiumcodereview.appspot.com/12089048