Telemetry: Don't require strict host key checking for CrOS
As part of getting telemetry into the autotest lab we need to disable
strict host key checking on all ssh and scp operations with the DUT.

The reason for this is that the DUT's in the autotest lab are constantly
being reimaged, thus making them look like different hosts as the same ip.

In order to get around this, we disable the StrickHostKeyChecking option,
and create a temporary file to act as our known hosts file. And on the
intial login, we clean up any Warnings that may occur so that we can
create the initial connection.

BUG=chromium:170462
TEST=Powerwashed my CrOS test machine, then ran telemetry with the changes
     and ensured that it worked as expected.


Review URL: https://chromiumcodereview.appspot.com/12077007