views: Remove simple_message_box_win.cc in favor of simple_message_box_views.cc
simple_message_box_win.cc is not be needed nor compatible with the desktop-aura
transition. It also uses the win32 appearance not the new Chrome style
appearance currently in development. So it's bad twice ;)

Just kill it and that eases the work that will need to be done later.

By enabling simple_message_box_views on Windows means that it will be used for all
ShowMessageBox() dialogs instead of current Windows MessageBox.

BUG=166075,54988
R=ben@chromium.org

Review URL: https://codereview.chromium.org/12042083