Make object on stack have name to avoid early space re-use
In recent (R25/ToT) ChromeOS builds, we are seeing Hangouts crashes in O3D. I
traced the crashes down to a crash in libjpeg, but apparently caused by O3D
allocating a class on the stack that the libjpeg structs use during decompress.
While the stack in question is still in scope, I believe the toolchain is seeing
an unnamed class as something no longer needed, despite the pointers into it
from the libjpeg struct.

Simple fix is to name it, so it remains in scope throughout the method.

BUG=chromium-os:38371
TEST=emerge-lumpy o3d, run Hangouts and mute video camera
Review URL: https://codereview.chromium.org/12087079