[NaCl SDK] Remove old SDK bundles from manifest (<= 20).
Also remove gdb_builds.

update_nacl_manifest uses omahaproxy to determine the Chrome version to use
when building the NaCl SDK bundle, but it only keeps the last ~5 revisions. As
a result, old versions need to either be pinned or purged.

The old versions are obviously still available on commondatastorage, they just
are not as easy to access from "naclsdk update" -- (we have old manifests up on
commondatastorage as well, so an intrepid user could still get old versions if
need be.)

BUG=none
R=noelallen@chromium.org
NOTRY=true


Review URL: https://chromiumcodereview.appspot.com/12079082