Fix failing test on Mac 10.6: TestInstallThemeInFullScreen
The initial change in r179554 [1] presumed that IDC_FULLSCREEN would only be issued on Mac 10.7 and later. TestInstallThemeInFullScreen calls fullscreen without a platform & os version check.

This solution is to tolerate IDC_FULLSCREEN and convert it to a IDC_PRESENTATION_MODE on Mac 10.6 and earlier. Both are conceptually the same "fullscreen with no browser chrome".

[1] http://src.chromium.org/viewvc/chrome?view=rev&revision=179554

BUG=169138
TBR=sky@chromium.org


Review URL: https://chromiumcodereview.appspot.com/12096063