Switch do GDI rendering when the screensaver is active.
When the screen saver or login-screen are active, any calls we make to d3d present get
ignored. As a result, after coming back, the browser displays stale contents until the
page refreshes again. This is only a problem if AERO (the Windows compositor) is turned
on.

With AERO on, our window doesn't seem to be getting any useful notifications when we enter or exit the
screensaver that we could use to trigger a repaint. With AERO off we get a WM_PAINT which forces
us to repaint.

If we use the GDI display path (which requires an expensive texture readback and a blit)
the contents of the window update correctly. The unfortunate side-effect of this change is that
while in screen saver mode our CPU utilization for pages that update often (e.g. have an
active CSS animation) will significantly increase.

BUG=133121


Review URL: https://chromiumcodereview.appspot.com/11863013