MediaStreamCaptureIndicator crash fix for multiple removes of the same device.
Oddly enough, the crash only occurs on the Mac platform: A NULL dereference was being made, which could only indicate that code calling into MediaStreamCaptureIndicator is attempting to remove the same device(s) more than once.

Root cause of the multiple "remove" calls has not been determined, but it is outside the scope of changes made in revision 179490 which introduced the crashing.

BUG=173198
TBR=xians


Review URL: https://chromiumcodereview.appspot.com/12088082