Remove/simplify the POSIX SharedMemory::Create() a bit.
This gets called on the IO thread. On Mac, operations done on a real
(disk-backed) filesystem (not a RAM/swap-backed tmpfs), and are slow:
  - Use unlink() directly instead of file_util::Delete(). This saves a stat
    (which is used to determine if the target is a directory or not).
  - Don't bother doing the fseeko(). This is unneeded, since mmap() takes an
    absolute offset. (ftruncate() will grow files both on Linux and Mac. On
    non-XSI POSIX systems where this is not the case, the seek alone isn't
    sufficient to grow the file in any case.)
Both these show up on a profile of a big session restore on Mac (though
unfortunately the bulk of the time is spent in mkstemp() and its internal file
open). (There's still a redundant stat, perhaps to be taken care of later.)


Review URL: https://chromiumcodereview.appspot.com/12089054