Use uint64 instead of size_t in GpuMemoryManager
There are two reasons for this. Most pressing is that this is skating
near the edge of overflow (especially where we multiply things by 9/8).
Second is that it is indeed possible to have a Gpu with more than 4GB
of VRAM accessible even to a 32-bit process.

BUG=168413


Review URL: https://chromiumcodereview.appspot.com/12091051