Refactor: Simplify WaitableEventWatcher.
This change uses a callback instead of a delegate for specifying what
should be called when a WaitableEvent occurs.

This simplifies the class and gets rid of a workaround internal to the
class to prevent name collision on "Delegate" inner classes.

Tested (on linux and windows):
  ninja -C out/Debug chrome
  out/Debug/base_unittests --gtest_filter=*WaitableEventWatcherTest*

BUG=


Review URL: https://chromiumcodereview.appspot.com/11953112