First cut at UI for saving net_logs data into a temporary file on Mobile (Android and iOS).
Added a new page chrome://net-export to start/stop/send the net_log
for mobile platforms. net_log entries are saved in
"chrome-net-export-log.json" file created in temporary directory.

Button "Start Saving (after deleting old data)" will start saving to a
temporary file in "NetLog" directory in Temp directory. It deletes all
the old temporary files that were in that directory before creating a
new temporty file. It is a no-op if we are already saving data into a
file.

Button "Stop Saving Data" closes the temporary file so that it could
be mailed. It is a no-op if we haven't started saving data into a
file.

Button "Send Data" emails the data file if a temporary file exists in
"NetLog" directory. It is a no-op if there is no file.

There is only a single temporary file into which we log the
data. NetLog data is saved in the browser process. Logging of net_log
data behaves similar to the command line option --log-net-log.

If chrome://net-export is accessed in multiple tabs, action in first
tab wins. If Stop is clicked in any tab, logging stops. If start is
clicked in any tab, net logging starts.

about:about lists chrome://net-export as one of the links.

unit tests create a temporary file for each test run and they delete the
file at the end of the test run.

mmenke@ and/or eroman@ for chrome\browser\resources and webui changes. 
jar@ for net_log_temp_file changes.
jhawkins@chromium.org for chrome\browser\ui\webui and chrome\browser\resources
OWNERS stamp.

R=jar@chromium.org, mmenke@chromium.org
BUG=151212
TEST=chrome://net-export

Review URL: https://chromiumcodereview.appspot.com/11828036