Delete flag --enable-css-transform-pinch.
This flag activated a different coordinate space for pinch that was used
only on Android.  Android has now converted to desktop pinch and this
old mode is already starting to break, so we should delete this flag.
I'll follow up with further cleanup later.

I also set apply_page_scale_factor_in_compositor to true in all cases in
order to avoid needing any further support for the codepaths where it's
false (it should be a no-op when pinch is disabled in principle).

NOTRY=true
BUG=173534


Review URL: https://chromiumcodereview.appspot.com/12089100