cc: Don't do full tree sync unless needed with impl painting.
When impl painting is enabled, we always did a full tree sync from
the main thread to the pending tree because it was always recreated.
Since we recycle the tree now, it maintains its structure from one
commit to the next, so we don't need to do a full tree sync unless
its structure changed, the same decision as for non-impl-painting
mode.

On the impl side, the pending tree maintains the same structure as the
active tree unless the main thread commits a structure change. This is
the case when the main thread does a full tree sync into the pending
tree. So we propagate the flag to the LayerTreeHostImpl and use it when
activating the pending tree.

Causing a commit every layout() with the 10_10_layer_tree, the win is clear:
Before: *RESULT 10_10_layer_tree: frames= 2657.12 runs/s
After: *RESULT 10_10_layer_tree: frames= 3256.78 runs/s

R=jamesr,enne
BUG=173526


Review URL: https://chromiumcodereview.appspot.com/12094094