Fix shared memory leak on command buffer teardown
When freeing mapped memory from the GLES2Implementation, it's pending token so
we don't destroy the TranserBuffer. When we tear down the GLES2Implementation,
there's still usually some TransferBuffers left, that never get explicitly destroyed in
the command buffer. When the command buffer gets destroyed, it deletes the
SharedMemory (which unmaps / closes), but doesn't tell the GPU process. In the
GPU process, the TransferBuffers are managed per ContextGroup, not context, so
they leak until the ContextGroup is destroyed.
It can happen e.g. when closing pages that share a renderer with another one, or
on Aura when creating/destroying root windows (on Chrome OS when unplugging a
screen).

BUG=chrome-os-partner:17496


Review URL: https://chromiumcodereview.appspot.com/12087126