LoadTiming int net 6:  Cache LoadTimingInfo
LoadTiming information is now cached at the URLRequest
level, so it can be read at any point after the headers
are received, until/unless the request is restarted.

The old behavior of keeping the information only as long as
it could be retrieved resulted in some problems.
Review URL: https://codereview.chromium.org/12114052