Relanding "bookmarks: Make bookmark manager paste at the right location."
It was reverted in r177400, due to break ExtensionApiTest.BookmarkManager.

Pasting a Bookmark page from main folder into a subfolder was not working.

To reproduce, one can do this:

- Create a bookmark page in bookmarks bar.
- Create a folder
- Copy or cut that bookmark page.
- Try to paste into that folder.

What we see today is that, instead of pasting the bookmark into the folder,
it pastes into the same folder but one index (+1) further.

With this patch, what we see is that the bookmark is pasted into the subfolder,
which is the expected behavior.

BUG=170296
TEST=follow the instructions above, there is also a good video showing how to
reproduce this bug.
R=sky@chromium.org,asargent@chromium.org,dbeam@chromium.org

Review URL: https://codereview.chromium.org/11931036