Check the length of a filename in ChromeOS "save as" dialog.
In the save-as dialog, it should be considered as a bug to return
a path that cannot be used to save a file. Otherwise we have to
check it everywhere we use the file chooser dialog.
We're already verifying, for example, invalid characters (like '/') but not the length.

Especially on ChromeOS, we have 143-byte limit under homedir and
the bar is low compared to other platforms, it became more problematic.

BUG=162734


Review URL: https://chromiumcodereview.appspot.com/12051035