Telemetry: Make ListAllDepsPaths pull DEPS locally
Currently ListAllDepsPaths goes to the internet to pull down the DEPS
required to run telemetry. This will give the caller the DEPS currently
on ToT and the use case is so that it pulls the DEPS from the files
around it in its checkout.

Change is to rework deps_include to be a dict, and to open the local DEPS
files  rather then grab them from the web.

BUG=chromium:173652
TEST=Ran ListAllDepsPath in python interpreter.
NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/12082118