Prevent Repeated Enter Keypresses when Selecting Autofill Entry.
This patch fixes a bug on the Windows implementation of the new Autofill
UI where if the enter key was used to select an entry, the entry was selected
and the form was also submitted (from the same key press).

BUG=
TEST=Using enter to select a Autofill option results in that entry getting
selected, without the form also getting submitted.


Review URL: https://chromiumcodereview.appspot.com/12045010