Merge 180288 - Fix enable state of Pepper Flash.
Component-updated and bundled Pepper Flash share the same enable state. When
reading from the Preferences file, the previous code ignored the enable state
if it was associated with the component-updated path. That was wrong because if
the component-updated Flash was newer, the enable state would be recorded with
that path. The outcome was that everytime users restarted the browser, Pepper
Flash was enabled, even if they disabled it before.

TBR=bauerb@chromium.org
BUG=150596
TEST=None

Review URL: https://codereview.chromium.org/12079044

TBR=yzshen@chromium.org
Review URL: https://codereview.chromium.org/12200006