History: Add range navigation control for grouped visits
The current patch enables the user to select the time frame
for which the the grouped results are viewed by
adding a set of buttons. These buttons allow the user to
go back in time, reset to today or go forward until the current interval.

BUG=170690

Review URL: https://chromiumcodereview.appspot.com/11886104