Implement DownloadFile operation on Drive API v2.
By this CL, DriveApiService supports downloading a file.
Unfortunately, the name of url for the downloading is different between
on GData WAPI, where it is "content url", and on Drive API v2, where it is
"download url". "Download URL" sounds a bit generic one, so this CL use it.

BUG=148627
TEST=Ran unit_tests.


Review URL: https://chromiumcodereview.appspot.com/12084108