Hardcode HOST_DESKTOP_TYPE_NATIVE all over the automation layer.
The automation layer was already hardcoding HOST_DESKTOP_TYPE_NATIVE all over the place and using BrowserList:: methods which were native desktop only; simply making that definitive, if we ever need the automation layer to work with multi-desktop envrionments we can come back later and fix it as a separate task (for now this contributes to getting rid of generic BrowserList:: methods).

Added a BrowserListImpl::get(int index); method to replace the *(...->begin() + index) paradigm which is completely unecessary given the underlying structure of a BrowserListImpl is a std::vector.

BUG=129187


Review URL: https://chromiumcodereview.appspot.com/12217041