[chromedriver] Fix hanging websocket tests and don't CHECK in tests.
The tests involving large messages were hanging on some machines because
net::StreamListenSocket::SendInternal was hanging waiting for socket to write,
which couldn't happen because the same blocked thread was also reading the
socket. Fixed by moving the server onto a separate thread. Also, moved
server into new TestHttpServer class, since a good portion was duplicated
between the sync and non-sync version of the tests.

Also, removed usage of CHECK in the tests and replaced with ASSERT, since
it's not a great to crash in tests.


Review URL: https://chromiumcodereview.appspot.com/12225028