Improve sorting performance with an explicit live-tile sift pass
This patch explicitly sifts live tiles to the front of the tiles_ list before calling std::sort only on the live tiles.  This preserves having exactly one list and all behaviors associated with it while significantly improving the performance of the sort.  The live tiles sub-section is about 10-20% the size of the whole list.

BUG=174202

Review URL: https://codereview.chromium.org/12210040