Mac: One click signin dialog
This is the Mac version of this Views CL:
https://codereview.chromium.org/11896021/

Currently the one click signin confirmation UI is implemented as a
bubble. This CL adds a tab modal dialog version of that UI.

Code common to the dialog and bubble has been moved to a new OneClickSigninViewController class.

XIB Changes:
  OneClickSigninBubble.xib:
    * aligned the baseline of the text in the advanced link with the
      text in the OK button.
    * changed the file owner to be OneClickSigninViewController
  OneClickSigninDialog.xib
    * same as OneClickSigninBubble.xib except it uses constrained
      window sytle controls

Screenshots:
  Bubble: http://i.imgur.com/HmJqoQk.png
  Dialog: http://i.imgur.com/gaY2cym.png

BUG=171330
TEST=
  To show the bubble I performed the following steps;
    - ran chrome with --use-client-login-signin-flow
    - signed into gmail
    - clicked "Ok Sync" in the infobar
  Testing the dialog requires access to server side stuff that's not public yet. I asked rogerta@ for instructions.

Review URL: https://codereview.chromium.org/12082088