[Spellcheck] Bring platform codepaths closer.
This is in preparation for async support on OSX - as a result of that,
codepaths across platforms will need to merge. As a first step, this
supports request caching "pro forma", without actually caching.

(The reason for that is the cache is for Red Underline only, and that
is still pending for OSX)

R=rlp@chromium.org
BUG=none


Review URL: https://chromiumcodereview.appspot.com/12189016