Implement focus based hit test logic
The WebKitClassic implementation of hit test based on focused node
change is incomplete, brittle, and racy. Not doing much better in this
patch since they are all busy design and effort required to improve it
is not worthwhile.

Traverse the WebKit render tree to get the relevant hit test information
for a focus node that is a link and possibly imbedding an image.

Note that the focused node becoming null (no node is focused), the
hit test results will not be nulled to avoid race with the touch path.

And what an oxymoronic title.

BUG=
Android only change. Ran through android bots.
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/12084114