Implement user consent InfoBar for screen capturing.
Support for screen capturing using getUserMedia() was added 
in r180502 (behind a flag), but there were no notifications 
to let the user know that the screen is being captured. This 
CL dds InfoBar that is shown whenever a page requests screen 
capturing and lets the user approve or deny the request.

BUG=134249
TBR=brettw@chromium.org (gypi changes)

Review URL: https://chromiumcodereview.appspot.com/12022024