Kill renderer processes that fail check in RenderViewHostImpl::SetWebUIProperty
When setting WebUI property for a renderer, it should have WebUI bindings enabled. If we get into inconsistent state where the renderer doesn't have those bindings, but we are trying to set WebUI properties, we should kill it. This is a defense in depth measure against security exploits.

BUG=174063

Review URL: https://chromiumcodereview.appspot.com/12210038