Omnibox: Create Keyword Verbatim Result in Search Provider
It's a little weird that SearchProvider creates the default provider's
search-what-you-typed (a.k.a. verbatim) match, the default provider's
suggestion matches, and the keyword provider's suggestions, but not
the keyword provider's search-what-you-typed match.

This change adds code to create the keyword-search-what-you-typed 
suggestion (for keywords that aren't extensions) to SearchProvider.  It
also modifies the code in KeywordProvider to prevent creating the
keyword-search-what-you-typed suggestions in these cases.

This is the natural step to fix the linked bug.  It also preps the path to get suggested relevance scores for keyword providers.

It should have nearly no user-visible effect.  The reason for the nearly no
and not a "no" is because now that the SearchProvider is generating
the keyword verbatim result, we allow the SearchProvider to return
an additional result to the AutocompleteController.  (The keyword provider
would always return this result, so we leave an extra slot in SearchProvider's
result set for this result.)  However, it's vaguely conceivable if we
have suggest relevance scoring and have more than four very-highly-scoring
suggestions that those suggestions could starve out the keyword verbatim
result.  Frankly, I'm not worried about this.

Note that there is a logging change: the verbatim result for the
keyword provider's search was previously marked as Keyword
provider, SEARCH_OTHER_ENGINE result type.  Now it will be
marked as Search provider, SEARCH_OTHER_ENGINE result type.

As for how I created CalculateRelevanceForKeywordVerbatim(), it's a direct
copy of KeywordProvider::CalculateRelevance() with certain tests removed
for things that are guaranteed to be true/false in SearchProvider if we're
using the keyword fetcher.  I decided not to make this code in common
because there's no reason the Keyword scoring algorithm should
be the same as the SearchProvider's scoring algorithm.  i.e., if one
later alters one of these functions, no harm will come as a result of
the fact that the code isn't shared.

BUG=171104

Review URL: https://chromiumcodereview.appspot.com/12090006