NaCl: Remove use of nacl::DescWrapper from nacl_ipc_adapter.cc
NaCl's DescWrapper class is deprecated and we don't want Chromium-side
code to be using it.  The main reason for using it before was the
Import*() methods, but we've added C constructors to replace these.

So that RewrittenMessage unrefs NaClDescs automatically, we add a new
wrapper class.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3278
TEST=NaCl tests in browser_tests


Review URL: https://chromiumcodereview.appspot.com/12217023