Gamepad: Linux support for DragonRise Generic USB.
This adds gamepad support for the DragonRise "Generic USB" controller, which is
supported on Mac, but wasn't previously on Linux.

Tested:
Dave had one of these controllers on hand, so it was easy to verify the correct
behavior. I used the Gamepad example in the NaCl SDK and checked that all
buttons and the axes worked properly.

BUG=


Review URL: https://chromiumcodereview.appspot.com/12223035