Turn off the network proxy for mac and win in DRT
The original test_shell code was written to use the system proxy
autoconfig. On windows, at least, this was known to be flaky at best.
However, this still made sense because we needed to be able to use
test_shell to view arbitrary pages on the web, and we didn't have an
alternative.

However, we don't really use test_shell anymore, and the same code
is used by DumpRenderTree, which only views pages on the local machine
and has no need for the proxy at all. This patch changes the net config
to never use a proxy on any platform.

R=tony@chromium.org, eroman@chromium.org
BUG=170429

Review URL: https://codereview.chromium.org/11938006