cc: Reset rasterScale to 1 when it changes.
If a webpage is changing the CSS scale of an element, we can't predict
what scales it is going to end up at, and whether it's going to look
good or end up being blurry. Whenever the scale changes (not inside
an accelerated animation), we reset the rasterScale for the layer to 1.0
and save it there.

This is meant to fix blurry composited layers when javascript-based
animations are used, without regressing static content. This problem
affects M24 through to trunk.

This can introduce an extra invalidation, but limits that to once for
the lifetime of the layer. Since we can't predict the static vs dynamic
case from the compositor, I feel this is a reasonable solution
until impl-side painting.

Tests:
LayerTreeHostCommonTest.verifyContentsScale

R=enne,jamesr
BUG=160218

Review URL: https://chromiumcodereview.appspot.com/12212021