Put a placeholder app in the app list during install.
This change adds a new item to the app list as soon as an app starts
installing, rather than waiting for the install to finish. Currently no icon is
displayed for it (it's blank except for the title), but the follow-up patches
will make it use the app's icon with a progress bar underneath.

Original review is at https://codereview.chromium.org/12047109/ but I had to re-upload due to a bad base URL.

BUG=152854
TBR=xiyuan@chromium.org,benwells@chromium.org

Review URL: https://codereview.chromium.org/12209046