Allow Terms of Service to be specified for Public Accounts
This CL allows Terms of Service to be specified for a Public Account. The
terms are set by specifying a URL through policy. Chrome downloads the
terms and presents them to the user. Only if the user accepts the terms
can the Public Account be entered.

This CL also fixes the regression described in issue 174314.

TEST=Manual
BUG=chromium-os:32780,174314

Review URL: https://codereview.chromium.org/12218014