Skip loading webgl test expectations on win gpu bots.
So we skip the tests in previous CL, and the random crashes still happen.  So it's apparent something else is causing it.

Try to see if loading the file is the cause of unexpected random crashes.

This is temporary and will be reverted once investigation is finished.

BUG=165269
TEST=win gpu bots
TBR=kbr
Review URL: https://codereview.chromium.org/12217092