Move event generation from LayerAnimationController::animate to a new method
This moves event generation from LayerAnimationController::animate to a
new method LayerAnimationController::updateState. With this change, we
can avoid skipping over the initial portion of an animation as a result of
skipping frames: we only call updateState after we know we haven't
skipped the frame.

BUG=123593


Review URL: https://chromiumcodereview.appspot.com/12221079