Fix dependencies.
Change the way directory dependencies are handled.  This allows us to create a directory:
  $(TOOLCHAIN)/$(CONFIG)/foo
in the case the source file is named
  foo/bar.c

The CL also improves how we handle dependencies of libraries.  Now the LIB macro understands a STAMP target.  The parent make can call and inject a STAMP target which
will only get touched if the library is rebuilt.  This
fixes the "always link & run NMF" we've been seeing in
the SDK.

TBR=sbc@chromium.org
NOTRY=true

Review URL: https://codereview.chromium.org/12210066