isolate.py: Allow absolute PATH_VARIABLES and symlinks pointing outside the build root
The ChromiumOS build system passes an absolute PRODUCT_DIR path (which points inside the build tree) and symlinks the product directory itself to an absolute path that is outside the build tree. This change causes isolate.py to accept absolute path variables as long as they point inside the tree, and to treat symlinks outside the build root as ordinary directories (or files).


BUG=None
TEST=tests/*, output of all Chrome _run targets unchanged

Review URL: https://chromiumcodereview.appspot.com/12211082