Fix issue 174250, crash when reloading packaged app.
When a shell window was closed, it wasn't removed from the registry immediately
(it was removed in a callback from OnNativeClose, which happens a bit later.)
This meant that when the app was loaded again, the window was still in the
registry with a stale pointer to the old extension. ExtensionSettingsHandler
was tickling this by iterating shell windows on extension load.

BUG=174250

Review URL: https://codereview.chromium.org/12210107