Make devtools_window.cc multi-desktop aware.
Getting rid of BrowserList:: methods which only know about the native-desktop.

Note that CreateBrowserForDevTools is still native desktop only so that devtools will not fully work yet in a multi-desktop environment, but this is making progress in the refactoring goal of getting rid of all native desktop only methods.

BUG=129187


Review URL: https://chromiumcodereview.appspot.com/12210063