cc: Fix PostiveRatio crash by checking raster scale
The first time through ManageTilings, raster scale is zero, so you can't
use it to find a ratio.  This DCHECKS because the ratio is infinite.
If the raster scale is zero, we already create new tilings, so skipping
this check will just avoid the DCHECK.

R=reveman@chromium.org
BUG=174236


Review URL: https://chromiumcodereview.appspot.com/12209103