Replace the misspelled range within a selection to correct spelling.
This CL changes the spell check correction to replace the misspelled range
within a selection instead of replacing the whole selection. This change enables
spell check of words selected with surrounding whitespace or punctuation. Such
selection happens on a double-click of a word with a trailing underscore in
Windows, for example. This CL requires the change in https://webkit.org/b/108513.

BUG=167485,169659

Review URL: https://chromiumcodereview.appspot.com/12038037