Enable compositor frame message.
The cause of the flaky tests this was causing is now better-understood
and we can re-enable this (the flaky tests are still disabled).

Note that this flag is also enabled by default on phones from Java,
so we should turn it on here to get the behavior everywhere and
stop needing to support the old path.

NOTRY=true
BUG=161945


Review URL: https://chromiumcodereview.appspot.com/12210132