Introduce dynamic ablation of AutocompleteProviders using AC dynamic trials.
The idea is simple. Use Autocomplete dynamic field trial group names to indicate which providers should be disabled. This information, due to limitations in field trial design, is passed in the group name in a string form of "DisabledProviders_<bitmask>".

For instance, in order to run a SHORTCUTS ablation, one need to create a field trial with a group name that contains "DisabledProviders_512" (1<<9).

BUG=174694
TBR=sky@chromium.org

Review URL: https://chromiumcodereview.appspot.com/12220004