Added deprecation notice for web intents.
Also added public template for app_intents.html doc back in so that previous bookmarks do not result in a 404.

Note that I've repeated the deprecation warning under each <h2> in case a developer winds up in a section via search. Eventually we will want to add an 'Instead, use...' sentence/paragraph once this is defined and at the same time
remove the bulk of content in place.

BUG=174262

Review URL: https://codereview.chromium.org/12211131