Sync FileSystem: disable etag match on duplication recovery.
This causes a PRECONDITION error when two clients try to delete duplicated files.
The conflict is ignorable on this case, so we can just drop the etag matching. 

(notry, since most of try jobs are passed, and failed try looks not related to the patch)

BUG=175682
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/12209115