Sync FileSystem: Update metadata database even when the file is not modified.
For current impl, Resource ID of a file is not updated for this sequence:
 (1) Add a file with resource_id_1
-- wait for synced --
 (2) Delete the file
 (3) Add a file with resource_id_2, and same content with the first file.

If the (2) and (3) come together, (2) is squashed into (3). And then (3) becomes nop since the content is not modified.

As the result, the metadata of the file has wrong Resource ID and all other change is ignored.

(notry, since most of try jobs are passed, and failed try looks not related to the patch)

BUG=175667
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/12223088