Removing the menu entry to start the ash desktop in Windows.
Previouly we removed it from Win8 but it is really just something
that is experimental and now that Aura is deployed we don't want
the casual user to play with it. Accelerator entry is gone as well.

The menu entry is kept in debug mode

BUG=175036
TEST=menu entry is gone in canary, see bug
Review URL: https://codereview.chromium.org/12251009