Enlarge clickable area of the Files.app's butter bar's close action (x).
The previous (x) button was very tiny - 16x16 pixels, which may be too small when using a touchpad.
This patch modifies css so the x button's clickable area is extended to the surrounding paddings. As a result, it is much easier to cancel an operation.
Along the way, simplified css for actions.

TEST=Copy a large file and try to cancel. Also delete to check if Undo link is not broken.
BUG=175950

Review URL: https://codereview.chromium.org/12218135