Add an API to component_updater that asks to do an update check "now".
With this API, users of component_updater can trigger an
update or install when deemed important.  E.g., if a
component is not yet installed, but is about to be used.

There is not a guarantee that the item will actually
be updated.  There is only a good chance.  It can
still be preempted by other items that may have an update
available as well, or if multiple components
use this interface one will trump the others.

Use this to kick off a check of PNaCl if flag enabled, but
not yet installed.

Sort of related to:
BUG=107438

Review URL: https://codereview.chromium.org/12054003