cc: Limit the total number of uploads.
We can't reference-count shared memory regions in the GPU
process, so we dup/map the memory block for each upload.
This doesn't scale for small uploads, both because we can
use too many file descriptors, and use too much address
space when the small upload transfer buffers are contained
in larger shared memory blocks.

This limits the total number of uploads to prevent the
problem, but this should be fixed properly and this limit
removed.

BUG=176002

NOTRY=True
Since it's a CC only change and unit tests have passed.


Review URL: https://chromiumcodereview.appspot.com/12260033