Fix visibility in compositing.
When visibility of the browser plugin is toggled on and off, the image remained on screen in the compositing mode.

This was introduced by the gutter layer used during resizing and autosize. The gutter layer is the compositing layer used for the browser plugin container. The content layer with the actual texture is the child of that gutter layer. When the container visibility is toggled, only the gutter layer was hidden. 

The solution is to plumb visibility changed to the compositing helper and hide the texture layer manually.

BUG=175752


Review URL: https://chromiumcodereview.appspot.com/12258026