Shill: Delay the ConnectToSignal call until an observer is added.
This patch delayes the call to ObjectProxy::ConnectToSignal made from the
ShillClientHelper in order to call it only when there's an observer for the
signal. If no observer is currently added, as in the case of a ShillClientHelper
used only for method calls, this fix will efectively cause the helper to not
add an unnecessary match rule in the dbus connection. This notably reduces the
number of match rules used by Shill to only the number of connected networks
instead of the number of available networks.

BUG=chromium:170182


Review URL: https://chromiumcodereview.appspot.com/12220128