Remove legacy task wrapper in MultiThreadedProxyResolver.
In the dark days before the grand unification of callback systems, there
was NewRunnableMethod(). It required a wrapper if the method you wanted
to bind to was on a non-refcounted type.

With base::Bind(), these wrappers are no longer necessary and the
wrapper simply hid the fact that we have an unretained reference to
ProxyResolver.

BUG=none

Review URL: https://codereview.chromium.org/12252022