Unify browser- and renderer- side logic for new DNS error pages
Create a new NetErrorTracker (TODO: better name?) that can be used on the browser and renderer side to watch the Start/Commit/Fail/Finish events and tell whether a tab or renderer is currently displaying a DNS error page.

This will be used immediately for the new NetErrorHelper (the renderer-side RenderViewObserver that receives the NetErrorInfo IPC from the browser-side NetErrorTabHelper); in a later CL, I'll convert NetErrorTabHelper to use it as well.

BUG=156415


Review URL: https://chromiumcodereview.appspot.com/12223098