cc: Changed Now to HighResNow in tile manager timings
When performing timing in tile manager (with --enable-gpu-benchmarking enabled), we want to ensure that we're using a high resolution clock. On non-windows platforms, the change ends up calling the same function. However, it is at least more descriptive of the intent of these calls.

BUG=173498


Review URL: https://chromiumcodereview.appspot.com/12194028