Fix select/unselect all for android IME.
This is used by context action modes as well as third
party keyboards and we need to ensure selection updates
are dispatched while executing either command in webkit.

Also fixes an issue where dismissing the contextual action
bar based on a selection change was clearing the selection.
This manifested itself by when you select a word in a text
field (the paste/copy/select all bar at the top would appear)
but then if you tap in the same text field it would notify
java of a selection bounds change and unselect everything
and hide the keyboard.  When the selection bounds change,
we just want to have the action bar hide without clearing the
selection as we are already responding to a selection update.

BUG=174939


Review URL: https://chromiumcodereview.appspot.com/12212152