Adding raw tracing to trace framework.
With some types of tracing you do not want to use the system time as the recorded event time. In those cases, we want to provide a timestamp that will be used to display the event.

In a similar vein, the thread which is outputting the event may not be the thread where the event happened so we also allow an explicit thread_id to be provided.

BUG=111509

Review URL: https://chromiumcodereview.appspot.com/11366109