Refactored FakeURLFetcher to make it more flexible
Added an optional parameter to the FakeURLFetcherFactory class that is a callback used to create the FakeURLFetcher. This can be used to add information to the FakeURLFetcher when it is created, such as HTTP headers or cookies.

As part of an effort to fix https://code.google.com/p/chromium/issues/detail?id=176361 .

BUG=

Review URL: https://chromiumcodereview.appspot.com/12211076