Make the backup/restore code in the test harness more fit.
Previously the backup/restore code would repeat code
to handle each of the paths.  This refactors the code
so that we reuse the same code for each of the paths
we want to backup and restore.

Also, this change runs the input paths through glob.glob
so we can pass in path inputs that use glob characters.

R=xusydoc@chromium.org


Review URL: https://chromiumcodereview.appspot.com/12253051