Use a uniform opaque type to represent web contents modal dialogs
Since the WebContentsModalDialog interface is going away, the
platform-independent code in WebContentsModalDialogManager requires an
opaque type with which to identify dialogs.

In a first attempt at this under https://codereview.chromium.org/12224020,
comments indicated that the Mac implementation of the web contents
modal dialog does not fit any of the abstracted native types provided
by ui/gfx/native_widget_types.h.  So, define a special typedef
NativeWebContentsModalDialog that uses the existing gfx::NativeView
abstraction on platforms where supported, mainly to document the expected
type. On Mac, we use a void* which is the preferred type for this, per
Sailesh in comments in this review.

BUG=157161


Review URL: https://chromiumcodereview.appspot.com/12221055