touch: Fix a crash in compositor-thread hit-test related code.
The browser makes an assumption that the compositor thread hit-testing will
never return NO_CONSUMER_EXISTS for a touch-event sequence if a previous event
in the same sequence was CONSUMED. This is not true, especially for multi-touch
touch sequences. So remove the check that enforces this assumption.

BUG=174988

Review URL: https://codereview.chromium.org/12282008