[cc] Sort tiles based on distance to screen after time, and before content rect
When the page is stationary, we break ties based on the content rect coordinates of the tiles. This causes us to basically ignore the worldspace importance of the tile completely! This means that when the viewport stops moving, and you have anything with more than one layer, far far far offscreen layers suddenly start getting memory.

Big thanks to @vangelis for figuring this out.


BUG=176645


Review URL: https://chromiumcodereview.appspot.com/12287023