Logs how often users clear Omnibox.
We're interested in the scenario that, a user wanted to use IME in Omnibox, but forgot to turn IME on, or opposite (wanted to not use IME, but forgot to turn IME off).  In these cases, most users realize their mistake just after typing a few characters, which are not what they expected, then delete all the text in Omnibox with backspace key (or escape key), turn IME on/off, and type again what they wanted to input.
We're interested in how often this scenario happens, and would like to see this frequency will reduce as we improve the Omnibox UX for IME users.

This CL adds a histogram named "Omnibox.UserTextCleared" which shows the number of times that users deleted all their input and the number of times that users hit escape key to revert their input.

BUG=172180
TEST=no regression

Review URL: https://chromiumcodereview.appspot.com/12211043