Delay the creation of the WebView until after the construction of the RenderViewImpl.
During layout tests, we use a class derived from RenderViewImpl, and therefore
the vtable of the WebViewClient is not yet completely set up during
RenderViewImpl's constructor.

Also change the configuration of the TestRunner library such that it doesn't
try to access the WebView before the RenderViewImpl is completely constructed
and initialized.

BUG=111316
R=jam@chromium.org

Review URL: https://codereview.chromium.org/12296002